+++ spec/support/test_app_template.rb
@@ -18,6 +18,7 @@ end
 
 template "README.md", force: true
 copy_file "env", ".env"
+copy_file "config/environments/ci.rb"
 template "manifest.yml"
 directory "doc"
 copy_file "githooks/pre-commit", ".githooks/pre-commit", mode: :preserve